# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual World 3D API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Position {
  x: number;
  z: number;
}

export interface Npc {
  id: string;
  name: string;
  color: string;
  position: Position;
  emotion: string;
  personality: string;
  currentAction: string;
  isMoving?: boolean;
  /** @nullable */
  memoryCount?: number | null;
}

export interface WorldObject {
  id: string;
  creator: string;
  creatorId: string;
  creatorColor: string;
  type: string;
  description: string;
  position: Position;
  createdAt: number;
}

export interface WorldState {
  npcs: Npc[];
  worldObjects: WorldObject[];
  totalPlayers: number;
  totalConversations: number;
}

export interface NpcMemory {
  id: string;
  npcId: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface NpcStats {
  npcId: string;
  name: string;
  totalConversations: number;
  totalCreations: number;
  mostRecentEmotion: string;
}