# lib/api-zod/src/generated/types/npc.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual World 3D API
 * OpenAPI spec version: 0.1.0
 */
import type { Position } from './position';

export interface Npc {
  id: string;
  name: string;
  color: string;
  position: Position;
  emotion: string;
  personality: string;
  currentAction: string;
  isMoving?: boolean;
  /** @nullable */
  memoryCount?: number | null;
}