# lib/api-zod/src/generated/types/npcMemory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual World 3D API
 * OpenAPI spec version: 0.1.0
 */

export interface NpcMemory {
  id: string;
  npcId: string;
  role: string;
  content: string;
  createdAt: string;
}