# lib/api-zod/src/generated/types/npcStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual World 3D API
 * OpenAPI spec version: 0.1.0
 */

export interface NpcStats {
  npcId: string;
  name: string;
  totalConversations: number;
  totalCreations: number;
  mostRecentEmotion: string;
}