# lib/api-zod/src/generated/types/worldObject.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual World 3D API
 * OpenAPI spec version: 0.1.0
 */
import type { Position } from './position';

export interface WorldObject {
  id: string;
  creator: string;
  creatorId: string;
  creatorColor: string;
  type: string;
  description: string;
  position: Position;
  createdAt: number;
}