# lib/api-zod/src/generated/types/worldState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual World 3D API
 * OpenAPI spec version: 0.1.0
 */
import type { Npc } from './npc';
import type { WorldObject } from './worldObject';

export interface WorldState {
  npcs: Npc[];
  worldObjects: WorldObject[];
  totalPlayers: number;
  totalConversations: number;
}